% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_coding.R
\name{binary_coding.prepare}
\alias{binary_coding.prepare}
\title{binary_coding.prepare - collect unique values form data columns in order to establish binary coding}
\usage{
binary_coding.prepare(data, columnIndices)
}
\arguments{
\item{columnIndices}{- columns selected for transformation}
}
\value{
- list of unique values per column
}
\description{
binary_coding.prepare - collect unique values form data columns in order to establish binary coding
}
