% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretization.apply}
\alias{discretization.apply}
\title{discretization.apply - apply discreatization intervals to data}
\usage{
discretization.apply(data, thresholds)
}
\arguments{
\item{thresholds}{- values computed by discretization.prepare function}
}
\value{
data with continuous values replaced with discrete values
}
\description{
discretization.apply - apply discreatization intervals to data
}
\examples{
 t = discretization.prepare(iris, c(1,2), list(list(5, 'even'),list(6, 'size')))
 transformed = discretization.apply(iris, t)
}
