% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{filter_outliners.apply}
\alias{filter_outliners.apply}
\title{filter_outliners.apply - removes rows where one of values in outliners columns meets filtering criteria}
\usage{
filter_outliners.apply(data, beta, quadrils)
}
\arguments{
\item{beta}{- smaller beta - more agresive filtering}

\item{quadrils}{obtained from filter_outliners.prepare function}
}
\description{
filter_outliners.apply - removes rows where one of values in outliners columns meets filtering criteria
}
\examples{
q <- filter_outliners.prepare(iris, columnIndices = c(1,2,3,4))
l <- filter_outliners.apply(iris, 0.1, q)
irisf <- iris[l,]
}
