% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation.apply}
\alias{imputation.apply}
\title{imputation.apply - fill data gaps (NA valuse) with values computed by inputation.prepare}
\usage{
imputation.apply(data, transformation_parameters, mark_artificial_vals = F)
}
\arguments{
\item{transformation_parameters}{- paramets obtainded from imputation.prepare function}

\item{mark_artificial_vals}{- for all columns with missing values add new column and mark artificial values in it (default False)}
}
\value{
data with NA values set based on transformation_parameters
}
\description{
imputation.apply - fill data gaps (NA valuse) with values computed by inputation.prepare
}
\examples{
iris_c = iris
fil = seq(from=1, to=150, by=2)
iris_c[fil, 1:2] = NA
p = inputation.prepare(iris_c, c(1, 2),  c("average", "median"))
after1 = inputation.apply(iris_c, p)
}
