% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation.prepare}
\alias{imputation.prepare}
\title{imputation.prepare - compute mean, median or mode for selected columns}
\usage{
imputation.prepare(data, columnIndices, columnOptions)
}
\arguments{
\item{columnIndices}{- select columns you want to process}

\item{columnOptions}{- select type of data inputation schema per column. One of:
- 'average'
- 'median'
- 'mode'}
}
\description{
imputation.prepare - compute mean, median or mode for selected columns
}
\examples{
iris_c = iris
fil = seq(from=1, to=150, by=2)
iris_c[fil, 1:2] = NA
iris_c[fil, 5] = NA
p = inputation.prepare(iris_c, c(1, 2, 5),  c("average", "median", "mode"))
}
