% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standarization.apply}
\alias{standarization.apply}
\title{standarization.apply - standarize data in selected columns}
\usage{
standarization.apply(data, transformation_parameters)
}
\arguments{
\item{transformation_parameters}{- paramets obtained from prepare_transformation}
}
\value{
data with column values replaced by new ones with mean = 0 and std = 1
}
\description{
standarization.apply - standarize data in selected columns
}
\examples{
p = prepare_transformation(iris, c(1))
r = apply_transformation(iris, p)
mean(r[, 1]) -> 0
sd(r[, 1]) -> 1
}
