% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standarization.prepare}
\alias{standarization.prepare}
\title{standarization.prepare - compute mean and std for columns}
\usage{
standarization.prepare(data, columnIndices)
}
\arguments{
\item{columnIndices}{}
}
\value{
list of (average, std) values per column selected in columnIndices
}
\description{
standarization.prepare - compute mean and std for columns
}
\examples{
p = standarization.prepare(iris, c(1))
}
