% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_coding.R
\name{binary_coding.apply}
\alias{binary_coding.apply}
\title{binary_coding.apply - add new binary values columns to data}
\usage{
binary_coding.apply(data, transformation_parameters)
}
\arguments{
\item{transformation_parameters}{- paramets obtained from prepare_transformation}
}
\value{
data with added binary columns
}
\description{
binary_coding.apply - add new binary values columns to data
}
\examples{
t = binary_coding.prepare(iris, c(5))
binary_coding.apply(iris, t) -> add 2 new colums binary representing 3 classes
}
